I added the maximize. minimize, buttons to the File dialog window. I changed the luminosity value type from integer to float.
git-svn-id: svn+ssh://svn.cern.ch/reps/dirac/LHCbDIRAC/trunk@29395 4525493e-7705-40b1-a816-d608a930855b

+++ LHCbDIRAC/BookkeepingSystem/Gui/Controler/ControlerFileDialog.py
@@ -228,7 +228,7 @@ class ControlerFileDialog(ControlerAbstract):
     for item in items:
       value = items[item]
       if value['TotalLuminosity'] != None:
-        luminosity += int(value['TotalLuminosity'])
+        luminosity += float(value['TotalLuminosity'])
     return luminosity
   
   #############################################################################
@@ -237,7 +237,7 @@ class ControlerFileDialog(ControlerAbstract):
     for item in items:
       value = items[item]
       if value['Luminosity'] != None:
-        luminosity += int(value['Luminosity'])
+        luminosity += float(value['Luminosity'])
     return luminosity
   
   #############################################################################
